Fix profile mobile tabs
